+++ src/umpire/op/MemoryOperationRegistry.cpp
@@ -50,6 +50,12 @@
 #include "umpire/op/SyclMemPrefetchOperation.hpp"
 #endif
 
+#if defined(UMPIRE_ENABLE_OPENMP_TARGET)
+#include <omp.h>
+#include "umpire/op/OpenMPTargetCopyOperation.hpp"
+#include "umpire/op/OpenMPTargetMemsetOperation.hpp"
+#endif
+
 #include "umpire/util/Macros.hpp"
 
 namespace umpire {
@@ -66,46 +72,46 @@ MemoryOperationRegistry::MemoryOperationRegistry() noexcept
 {
   registerOperation(
       "COPY",
-      std::make_pair(Platform::cpu, Platform::cpu),
+      std::make_pair(Platform::host, Platform::host),
       std::make_shared<HostCopyOperation>());
 
   registerOperation(
       "MEMSET",
-      std::make_pair(Platform::cpu, Platform::cpu),
+      std::make_pair(Platform::host, Platform::host),
       std::make_shared<HostMemsetOperation>());
 
   registerOperation(
       "REALLOCATE",
-      std::make_pair(Platform::cpu, Platform::cpu),
+      std::make_pair(Platform::host, Platform::host),
       std::make_shared<HostReallocateOperation>());
 
 #if defined(UMPIRE_ENABLE_NUMA)
   registerOperation(
       "MOVE",
-      std::make_pair(Platform::cpu, Platform::cpu),
+      std::make_pair(Platform::host, Platform::host),
       std::make_shared<NumaMoveOperation>());
 
 // NOTE: We don't use CUDA calls in the move operation so no guard is needed
   registerOperation(
       "MOVE",
-      std::make_pair(Platform::cpu, Platform::cuda),
+      std::make_pair(Platform::host, Platform::cuda),
       std::make_shared<NumaMoveOperation>());
 
   registerOperation(
       "MOVE",
-      std::make_pair(Platform::cuda, Platform::cpu),
+      std::make_pair(Platform::cuda, Platform::host),
       std::make_shared<NumaMoveOperation>());
 #endif
 
 #if defined(UMPIRE_ENABLE_CUDA)
   registerOperation(
       "COPY",
-      std::make_pair(Platform::cpu, Platform::cuda),
+      std::make_pair(Platform::host, Platform::cuda),
       std::make_shared<CudaCopyToOperation>());
 
   registerOperation(
       "COPY",
-      std::make_pair(Platform::cuda, Platform::cpu),
+      std::make_pair(Platform::cuda, Platform::host),
       std::make_shared<CudaCopyFromOperation>());
 
   registerOperation(
@@ -135,7 +141,7 @@ MemoryOperationRegistry::MemoryOperationRegistry() noexcept
 
   registerOperation(
       "PREFERRED_LOCATION",
-      std::make_pair(Platform::cpu, Platform::cpu),
+      std::make_pair(Platform::host, Platform::host),
       std::make_shared<CudaAdvisePreferredLocationOperation>());
 
   registerOperation(
@@ -155,7 +161,7 @@ MemoryOperationRegistry::MemoryOperationRegistry() noexcept
 
   registerOperation(
       "UNSET_PREFERRED_LOCATION",
-      std::make_pair(Platform::cpu, Platform::cpu),
+      std::make_pair(Platform::host, Platform::host),
       std::make_shared<CudaAdviseUnsetPreferredLocationOperation>());
 
   registerOperation(
@@ -172,12 +178,12 @@ MemoryOperationRegistry::MemoryOperationRegistry() noexcept
 #if defined(UMPIRE_ENABLE_HIP)
   registerOperation(
       "COPY",
-      std::make_pair(Platform::cpu, Platform::hip),
+      std::make_pair(Platform::host, Platform::hip),
       std::make_shared<HipCopyToOperation>());
 
   registerOperation(
       "COPY",
-      std::make_pair(Platform::hip, Platform::cpu),
+      std::make_pair(Platform::hip, Platform::host),
       std::make_shared<HipCopyFromOperation>());
 
   registerOperation(
@@ -194,7 +200,33 @@ MemoryOperationRegistry::MemoryOperationRegistry() noexcept
       "REALLOCATE",
       std::make_pair(Platform::hip, Platform::hip),
       std::make_shared<GenericReallocateOperation>());
+#endif
 
+#if defined(UMPIRE_ENABLE_OPENMP_TARGET)
+  registerOperation(
+      "COPY",
+      std::make_pair(Platform::host, Platform::omp_target),
+      std::make_shared<OpenMPTargetCopyOperation>());
+
+  registerOperation(
+      "COPY",
+      std::make_pair(Platform::omp_target, Platform::host),
+      std::make_shared<OpenMPTargetCopyOperation>());
+
+  registerOperation(
+      "COPY",
+      std::make_pair(Platform::omp_target, Platform::omp_target),
+      std::make_shared<OpenMPTargetCopyOperation>());
+
+  registerOperation(
+      "MEMSET",
+      std::make_pair(Platform::omp_target, Platform::omp_target),
+      std::make_shared<OpenMPTargetMemsetOperation>());
+
+  registerOperation(
+      "REALLOCATE",
+      std::make_pair(Platform::omp_target, Platform::omp_target),
+      std::make_shared<GenericReallocateOperation>());
 #endif
 
 #if defined(UMPIRE_ENABLE_SYCL)