+++ src/umpire/strategy/NumaPolicy.cpp
@@ -24,13 +24,13 @@ NumaPolicy::NumaPolicy(
     int numa_node):
   AllocationStrategy(name, id),
   m_allocator(allocator.getAllocationStrategy()),
-  m_platform(Platform::cpu),
+  m_platform(Platform::host),
   m_node(numa_node)
 {
   if (numa_node < 0) {
     UMPIRE_ERROR("NumaPolicy error: NUMA nodes are always non-negative ints");
   }
-  if (allocator.getPlatform() != Platform::cpu) {
+  if (allocator.getPlatform() != Platform::host) {
     UMPIRE_ERROR("NumaPolicy error: allocator is not of cpu type");
   }
 
@@ -66,18 +66,6 @@ void
 NumaPolicy::deallocate(void* ptr)
 {
   m_allocator->deallocate(ptr);
-}
-
-std::size_t
-NumaPolicy::getCurrentSize() const noexcept
-{
-  return m_allocator->getCurrentSize();
-}
-
-std::size_t
-NumaPolicy::getHighWatermark() const noexcept
-{
-  return m_allocator->getHighWatermark();
 }
 
 Platform